Move mathematical functions to utils.h
